feat: Add function to get terminal rows and columns dynamically
This commit adds a new function `getTerminalRowsAndCols` which uses the `ioctl` system call to get the current number of rows and columns of the terminal window. The function is called in the `main` function to set the global variables `ROWS` and `COLS` dynamically based on the size of the terminal window. This allows for proper rendering of the game board regardless of the size of the terminal window.